move DataDisplay to its own file (#1854)
* move DataDisplay to its own file

* [pre-commit.ci] auto fixes from pre-commit hooks

* fix accordion name

---------
