+++ containers/ecr-viewer/src/app/services/labsService.tsx
@@ -1,12 +1,6 @@
 import React from "react";
 import { Bundle, Observation, Organization, Reference } from "fhir/r4";
-import {
-  PathMappings,
-  DisplayDataProps,
-  DataDisplay,
-  ColumnInfoInput,
-  noData,
-} from "@/app/utils";
+import { PathMappings, ColumnInfoInput, noData } from "@/app/utils";
 import { evaluate } from "fhirpath";
 import { AccordionLabResults } from "@/app/view-data/components/AccordionLabResults";
 import {
@@ -20,6 +14,7 @@ import {
 import { ObservationComponent } from "fhir/r4b";
 import { evaluateTable } from "./evaluateService";
 import { evaluateReference, evaluateValue } from "./evaluateFhirDataService";
+import { DataDisplay, DisplayDataProps } from "@/app/DataDisplay";
 
 export interface LabReport {
   result: Array<Reference>;
