+++ containers/ecr-viewer/src/app/view-data/components/SocialHistory.tsx
@@ -1,4 +1,3 @@
-import { DataDisplay, DisplayDataProps } from "@/app/utils";
 import {
   AccordionSection,
   AccordionH4,
@@ -6,6 +5,7 @@ import {
 } from "../component-utils";
 import { SectionConfig } from "./SideNav";
 import React from "react";
+import { DataDisplay, DisplayDataProps } from "@/app/DataDisplay";
 
 interface SocialHistoryProps {
   socialData: DisplayDataProps[];
